+++ src/main/java/edu/umn/ecology/populus/model/ie/EquationCalculator.java
@@ -161,9 +161,9 @@ public class EquationCalculator implements Serializable {
                         }
                     }
                     if (tokenIdentified) {
-                        v[i].addElement(new Token(TokenEnum.kConstant, Double.valueOf(temp)));
+                        v[i].addElement(new Token(TokenEnum.kConstant, Double.valueOf(temp).doubleValue()));
                     } else {
-                        v[i].addElement(new Token(TokenEnum.kConstant, Double.valueOf(temp.substring(0, k - 1))));
+                        v[i].addElement(new Token(TokenEnum.kConstant, Double.valueOf(temp.substring(0, k - 1)).doubleValue()));
                         temp = temp.substring(k - 1, temp.length());
                     }
                 }
@@ -310,7 +310,7 @@ public class EquationCalculator implements Serializable {
                                         if (constants != null) {
                                             if (constants.containsKey(temp)) {
                                                 try {
-                                                    value = (Double) constants.get(temp);
+                                                    value = ((Double) constants.get(temp)).doubleValue();
                                                 } catch (NumberFormatException nfe) {
                                                     throw new IEException("Hashtable didn't give a number.");
                                                 }
@@ -643,11 +643,11 @@ public class EquationCalculator implements Serializable {
 
         //send the 2 halves back to this function.
         if (!foundSymbol) {
-            cs.push(right);
+            cs.push(Integer.valueOf(right));
         } else {
             InfixToPrefix(cs, index, op2, right);
             InfixToPrefix(cs, index, left, op1);
-            cs.push(symbol);
+            cs.push(Integer.valueOf(symbol));
         }
     }
 
@@ -823,13 +823,13 @@ public class EquationCalculator implements Serializable {
             }
             Vector a = GetVector(i);
             bf = new Stack();
-            while (((Token) a.elementAt((Integer) ix.peek())).tokenType != TokenEnum.kConstant || !bf.empty()) {
-                while (((Token) a.elementAt((Integer) ix.peek())).tokenType != TokenEnum.kConstant) {
+            while (((Token) a.elementAt(((Integer) ix.peek()).intValue())).tokenType != TokenEnum.kConstant || !bf.empty()) {
+                while (((Token) a.elementAt(((Integer) ix.peek()).intValue())).tokenType != TokenEnum.kConstant) {
                     bf.push(ix.pop());
                 }
-                if (((Token) a.elementAt((Integer) bf.peek())).tokenType == TokenEnum.kFunction) {
-                    sym = (Integer) bf.pop();
-                    op1 = (Integer) ix.peek();
+                if (((Token) a.elementAt(((Integer) bf.peek()).intValue())).tokenType == TokenEnum.kFunction) {
+                    sym = ((Integer) bf.pop()).intValue();
+                    op1 = ((Integer) ix.peek()).intValue();
                     ((Token) a.elementAt(op1)).value = PerformFunction((int) ((Token) a.elementAt(sym)).value, ((Token) a.elementAt(op1)).value);
                     while (!bf.empty()) {
                         ix.push(bf.pop());
@@ -837,10 +837,10 @@ public class EquationCalculator implements Serializable {
                     continue;
                 }
                 bf.push(ix.pop());
-                if (((Token) a.elementAt((Integer) ix.peek())).tokenType == TokenEnum.kConstant) {
-                    op1 = (Integer) bf.pop();
-                    sym = (Integer) bf.pop();
-                    op2 = (Integer) ix.peek();
+                if (((Token) a.elementAt(((Integer) ix.peek()).intValue())).tokenType == TokenEnum.kConstant) {
+                    op1 = ((Integer) bf.pop()).intValue();
+                    sym = ((Integer) bf.pop()).intValue();
+                    op2 = ((Integer) ix.peek()).intValue();
                     ((Token) a.elementAt(op2)).value = PerformOperator((int) ((Token) a.elementAt(sym)).value, ((Token) a.elementAt(op1)).value, ((Token) a.elementAt(op2)).value);
                     while (!bf.empty()) {
                         ix.push(bf.pop());
@@ -850,7 +850,7 @@ public class EquationCalculator implements Serializable {
                 }
             } //end while
             if (a.size() > 0 && used[i]) {
-                calculated[i] = ((Token) a.elementAt((Integer) ix.peek())).value;
+                calculated[i] = ((Token) a.elementAt(((Integer) ix.peek()).intValue())).value;
             }
         } //end for
     }
@@ -868,7 +868,7 @@ public class EquationCalculator implements Serializable {
             System.out.print("\n" + i + ").\t");
             System.out.print(l + "\t");
             for (int j = 0; j < l; j++) {
-                k = (Integer) temp.pop();
+                k = ((Integer) temp.pop()).intValue();
                 System.out.print(((Token) v[i].elementAt(k)).shortString());
             }
         }
